move first frame (MethodFrame) into separate file
